% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_from_R0.R
\name{calc_p_from_R0}
\alias{calc_p_from_R0}
\title{This function takes the input R0 and the parameter vector as inputs and calculates the
corresponding p}
\usage{
calc_p_from_R0(R0_input, vec)
}
\description{
This function takes the input R0 and the parameter vector as inputs and calculates the
corresponding p
}
