% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_R0_td_from_p.R
\name{calc_R0_td_from_p}
\alias{calc_R0_td_from_p}
\title{This function takes the input p and the parameter vector as inputs and calculates the
corresponding R0 and doubling time}
\usage{
calc_R0_td_from_p(params, p)
}
\description{
Calculate R0 and td from p
}
