% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes2.R
\name{compute_cases}
\alias{compute_cases}
\title{Compute Cases from Simulation Outcomes}
\usage{
compute_cases(out, hospitalization_rate = 0.17, ventilator_rate = 0.05)
}
\description{
Compute Cases from Simulation Outcomes
}
