% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{compute_cases_intervention}
\alias{compute_cases_intervention}
\title{Compute Cases from Simulation Outcomes - Intervention}
\usage{
compute_cases_intervention(out)
}
\description{
Compute Cases from Simulation Outcomes - Intervention
}
