% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes2.R
\name{compute_cumulative_cases_intervention}
\alias{compute_cumulative_cases_intervention}
\title{Compute Cumulative Cases from Simulation Outcomes - Intervention}
\usage{
compute_cumulative_cases_intervention(out, hospitalization_rate = 0.17,
  ventilation_rate = 0.05)
}
\description{
Compute Cumulative Cases from Simulation Outcomes - Intervention
}
