% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix_shiny_server_module.R
\name{contact_matrix_server_module}
\alias{contact_matrix_server_module}
\title{Contact Matrix Shiny Server Module}
\usage{
contact_matrix_server_module(input, output, session)
}
\description{
This Shiny module watches for when the user updates one of the 
contact matrix parameters and applies the appropriate updates to the other
}
