% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_epi.R
\docType{package}
\name{covid.epi-package}
\alias{covid.epi}
\alias{covid.epi-package}
\title{COVID19 Epi Model and Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alyssa Bilinski \email{abilinski@g.harvard.edu}

Authors:
\itemize{
  \item Christian Testa \email{ctesta@hsph.harvard.edu}
  \item Yu-Han Kao \email{yu-han.kao@yale.edu}
  \item Lin Zhu \email{linzhu1@stanford.edu}
  \item Kayoko Shioda \email{kayoko.shioda@yale.edu}
  \item Ruthie Bierger \email{ruthie.birger@yale.edu}
}

Other contributors:
\itemize{
  \item Joshua Salomon \email{salomon1@stanford.edu} [correspondent]
  \item Meagan Fitzpatrick \email{meagan.fitzpatrick@som.umaryland.edu} [correspondent]
}

}
