% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_diagram.R
\name{export_model_diagram_to_png}
\alias{export_model_diagram_to_png}
\title{Render the model diagram to a PNG file we can give users}
\usage{
export_model_diagram_to_png(output_file)
}
\description{
Render the model diagram to a PNG file we can give users
}
