% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{fit_model}
\alias{fit_model}
\title{Fit the Model}
\usage{
fit_model(params, observed_data)
}
\description{
Fit the Model
}
\examples{

  params <- load_parameters()
  observed_data <- filter_states_data('MA')
  fit_model(params, observed_data)
<<<<<<< HEAD
=======
  params$td <- 1 
>>>>>>> 97b124ea164ca586d68a6ffb52ce47eac3880f12


}
