% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{format_simulation_outcomes_for_plotting}
\alias{format_simulation_outcomes_for_plotting}
\title{Format Model Outcomes for Plotting}
\usage{
format_simulation_outcomes_for_plotting(sim_outcomes)
}
\description{
Format Model Outcomes for Plotting
}
