% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{format_simulation_outcomes_for_plotting_int}
\alias{format_simulation_outcomes_for_plotting_int}
\title{Format Model Outcomes for Plotting - Intervention}
\usage{
format_simulation_outcomes_for_plotting_int(sim_outcomes, sim_outcomes_int)
}
\description{
Format Model Outcomes for Plotting - Intervention
}
