% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{hosp_days_plots}
\alias{hosp_days_plots}
\title{calculate which day hospital beds exceeds count}
\usage{
hosp_days_plots(no_sd_mat, sd_mat, beds, num_scens = 6, hosp_time = 10,
  hosp_pc = 0.05, output_file)
}
\description{
calculate difference with and without SD
takes as inputs the outputs of run_vecs for scenarios with and without social distancing
}
\examples{

 # I wasn't sure how to get this working ... 

 # I tried different ways to call hosp_days_plots with outcomes from the simulation 
 # model and I couldn't quite figure out what I should do
 
 # Below I tried to make the data frame look like what I anticipated the function wants 
 # based on the variables it accesses, but I still couldn't get it to work.
 
  params <- load_parameters()
  params2 <- load_parameters(8) # SD 50\% scenario
  
  det_table <- load_detection_table()
  
  sim_out <- run_param_vec(params = params, days_out1 = 30, days_out2 = NULL,
    model_type = run_basic, params2 = NULL, det_table = det_table)

  sim_out_int <- run_param_vec(params = params, days_out1 = 15, days_out2 = 30,
    model_type = run_int, params2 = params2, det_table = det_table)

  sim_out <- format_simulation_outcomes_for_plotting(sim_out)
  sim_out_int <- format_simulation_outcomes_for_plotting(sim_out_int)

  sim_out <- compute_cases(sim_out)
  sim_out_int <- compute_cases(sim_out_int)

  sim_out \%<>\% group_by(time) \%>\% summarize(
    Detected = sum(Detected),
    Infected = sum(Infected),
    value = sum(Total))

  sim_out_int \%<>\% group_by(time) \%>\% summarize(
    Detected = sum(Detected),
    Infected = sum(Infected),
    value = sum(Total))

  sim_out$scenario <- 'base case'
  sim_out_int$scenario <- 'social distancing 50\%'
  
  # still throws an error
  hosp_days_plots(no_sd_mat = sim_out, sd_mat = sim_out_int, beds = 100)

}
