% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{hosp_time_plots}
\alias{hosp_time_plots}
\title{Plot hospital bed needs over time along with capacity}
\usage{
hosp_time_plots(out_formatted, beds, num_scens = 6, hosp_time = 10,
  hosp_pc = 0.05)
}
\description{
Plot hospital bed needs over time along with capacity
}
\examples{
 params <- load_parameters()
 det_table <- load_detection_table()
 sim_out <- run_param_vec(params = params, days_out1 = 30, days_out2 = NULL,
   model_type = run_basic, params2 = NULL, det_table = det_table)
 out_formatted <- format_simulation_outcomes_for_plotting(sim_out)
 hosp_time_plots(out_formatted, beds=100)

}
