% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Load Parameters Vector from Parameters Table}
\usage{
load_parameters(scenario = 1)
}
\arguments{
\item{scenario}{An integer level selecting one of the rows of the
parameters csv. 1 corresponds to the base case. Others correspond to
intervention scenarios.}
}
\description{
Load Parameters Vector from Parameters Table
}
