% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_time_series.R
\name{make_detection_table}
\alias{make_detection_table}
\title{Make Detection Table}
\usage{
make_detection_table(t, rdetecti, rdetecta)
}
\description{
Helper to make a detection table
}
\details{
Currently assumes constants over time
}
