% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{make_plots_int}
\alias{make_plots_int}
\title{Make Plots Comparing Base Case and Intervention}
\usage{
make_plots_int(test, params, test_int, params_int)
}
\description{
Make Plots Comparing Base Case and Intervention
}
