% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{model_loss_function}
\alias{model_loss_function}
\title{Our model penalty function is calculated as the sum of squared errors for the 
case series data provided}
\usage{
model_loss_function(params, observed_data)
}
\description{
Our model penalty function is calculated as the sum of squared errors for the 
case series data provided
}
