% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_strat}
\alias{model_strat}
\title{Stratified Model}
\usage{
model_strat(t, x, parms, parms_int, int_start_time, int_stop_time,
  det_input_method = "input", det_table = NULL)
}
\description{
Stratified Model
}
