% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_cases_by_symptom_status_int}
\alias{plot_cases_by_symptom_status_int}
\title{Plot cases by symptom status - Intervention}
\usage{
plot_cases_by_symptom_status_int(out)
}
\description{
Plot cases by symptom status - Intervention
}
