% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_cases_needing_advanced_care_int}
\alias{plot_cases_needing_advanced_care_int}
\title{Plot Cases Needing Advanced Care - Intervention}
\usage{
plot_cases_needing_advanced_care_int(out_cases, cumulative = TRUE)
}
\description{
Plot Cases Needing Advanced Care - Intervention
}
