% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_cumulative_cases_by_age}
\alias{plot_cumulative_cases_by_age}
\title{Plot Cumulative Cases by Age}
\usage{
plot_cumulative_cases_by_age(out_cases)
}
\description{
Plot Cumulative Cases by Age
}
