% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_cumulative_cases_by_age_int}
\alias{plot_cumulative_cases_by_age_int}
\title{Plot Cumulative Cases by Age - Intervention}
\usage{
plot_cumulative_cases_by_age_int(out_cases, cumulative = TRUE)
}
\description{
Plot Cumulative Cases by Age - Intervention
}
\examples{
# setup
param_vec <- load_parameters()
det_table <- make_detection_table(180, .1, 0)

param_vec_int <- param_vec
param_vec_int['s'] <- .5

# run simulations
sim_out <- run_param_vec(params = param_vec, days_out1 = 180, det_table=det_table)
# run intervention scenario
sim_out_int <- run_param_vec(params = param_vec, params2 = param_vec_int, 
   days_out1 = 30, days_out2 = 180, days_out3 = 180, det_table=det_table,
   model_type = run_int)

# reformat for plotting
sim_out_formatted <- format_simulation_outcomes_for_plotting_int(sim_out, sim_out_int)

# reformat for plotting cumulative cases
sim_out_formatted_cases <- compute_cases_intervention(sim_out_formatted)

# plot!
plot_cumulative_cases_by_age_int(sim_out_formatted_cases)
}
