% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_diagnosed_cumulative_cases_by_age}
\alias{plot_diagnosed_cumulative_cases_by_age}
\title{Plot Diagnosed Cumulative Cases by Age
Note that these are "observed" meaning diagnosed}
\usage{
plot_diagnosed_cumulative_cases_by_age(out_cases)
}
\description{
Plot Diagnosed Cumulative Cases by Age
Note that these are "observed" meaning diagnosed
}
