% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_fit_to_observed_deaths_data_int}
\alias{plot_fit_to_observed_deaths_data_int}
\title{Plot Fit to Observed Deaths Data - Intervention}
\usage{
plot_fit_to_observed_deaths_data_int(out,
  observed_data = load_SCC_time_series(), cumulative = TRUE)
}
\description{
Plot Fit to Observed Deaths Data - Intervention
}
