% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_flows_by_compartment2}
\alias{plot_flows_by_compartment2}
\title{Plot Flows by Compartment Version 2}
\usage{
plot_flows_by_compartment2(out)
}
\description{
Not sure what the difference between this and plot_flows_by_compartment is
}
