% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_flows_by_compartment2_int}
\alias{plot_flows_by_compartment2_int}
\title{Plot Flows by Compartment Version 2 - Intervention}
\usage{
plot_flows_by_compartment2_int(out)
}
\description{
Not sure what the difference between this and plot_flows_by_compartment is
}
