% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_flows_by_compartment_int}
\alias{plot_flows_by_compartment_int}
\title{Plot Flows by Compartment for Interventions}
\usage{
plot_flows_by_compartment_int(out)
}
\description{
Plot Flows by Compartment for Interventions
}
