% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_flows_by_compartment_strata}
\alias{plot_flows_by_compartment_strata}
\title{Plot Flows by Compartment - Broken down by Strata}
\usage{
plot_flows_by_compartment_strata(out)
}
\description{
Not sure what the difference between this and plot_flows_by_compartment is
}
