% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_flows_by_compartment_strata_int}
\alias{plot_flows_by_compartment_strata_int}
\title{Plot Flows by Compartment Version 2 - Intervention}
\usage{
plot_flows_by_compartment_strata_int(out,
  selected_compartments = c("Susceptible", "Exposed", "Asymptomatic",
  "Symptomatic", "Recovered"))
}
\description{
Not sure what the difference between this and plot_flows_by_compartment is
}
\examples{
# setup
param_vec <- load_parameters()
det_table <- make_detection_table(180, .1, 0)

param_vec_int <- param_vec
param_vec_int['s'] <- .5

# run simulations
sim_out <- run_param_vec(params = param_vec, days_out1 = 180, det_table=det_table)
# run intervention scenario
sim_out_int <- run_param_vec(params = param_vec, params2 = param_vec_int, 
   days_out1 = 30, days_out2 = 180, days_out3 = 180, det_table=det_table,
   model_type = run_int)

# reformat for plotting
sim_out_formatted <- format_simulation_outcomes_for_plotting_int(sim_out, sim_out_int)

# plot!
plot_flows_by_compartment_strata_int(sim_out_formatted)

}
