% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_diagram.R
\name{plot_model_diagram}
\alias{plot_model_diagram}
\title{Function that renders a model diagram}
\usage{
plot_model_diagram()
}
\description{
grViz from DiagrammeR returns an SVG diagram based on the user input provided.
}
\details{
It also provides a clean syntax to make sure that we will be able to update the 
model diagram easily in the future if needed.

This is useful for rendering in our Shiny app too, as an SVG is lighter than an 
image file and will never have compression artifacts.

Documentation for grViz here: 
http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html
}
