% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_ratio_of_new_to_existing_cases}
\alias{plot_ratio_of_new_to_existing_cases}
\title{Plot Ratio of New To Existing Cases}
\usage{
plot_ratio_of_new_to_existing_cases(out)
}
\description{
Plot Ratio of New To Existing Cases
}
