% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_ratio_of_new_to_existing_cases_int}
\alias{plot_ratio_of_new_to_existing_cases_int}
\title{Plot Ratio of New To Exisitng Cases - Intervention}
\usage{
plot_ratio_of_new_to_existing_cases_int(out)
}
\description{
Plot Ratio of New To Exisitng Cases - Intervention
}
