% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_basic}
\alias{run_basic}
\title{Run Basic Model}
\usage{
run_basic(model = model_strat, xstart, params = params,
  params2 = NULL, days_out1, days_out2 = NULL, days_out3 = NULL,
  det_table = det_table)
}
\description{
Run Basic Model
}
