% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_int}
\alias{run_int}
\title{Run Model with Intervention}
\usage{
run_int(model = model_strat, xstart, params = params, params2 = NULL,
  days_out1, days_out2)
}
\description{
Run Model with Intervention
}
