% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_model}
\alias{run_model}
\title{Run the Model}
\usage{
run_model(func, xstart, times, params, det_table, method = "lsodes",
  events = NULL, parms_int, int_start_time, int_stop_time)
}
\description{
Run the Model
}
