% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run the Interactive Projection Tool for COVID-19 Interventions Application}
\usage{
runApp()
}
\description{
Run the Interactive Projection Tool for COVID-19 Interventions Application
}
