% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_epi.R
\docType{package}
\name{covid.epi-package}
\alias{covid.epi}
\alias{covid.epi-package}
\title{COVID19 Epi Model and Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
