% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_epi.R
\docType{package}
\name{covid.epi-package}
\alias{covid.epi}
\alias{covid.epi-package}
\title{COVID19 Epi Model and Shiny App}
\description{
Provides an R package and Shiny web app interface to an epidemiological
    model of an SEIR-type model intended for use in modeling the transmission
    and outbreak of COVID-19 as well as transmission control scenarios.
}
\author{
\strong{Maintainer}: COVID-SPEC \email{c19spec@gmail.com}

}
