% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes2.R
\name{format_simulation_outcomes_for_plotting}
\alias{format_simulation_outcomes_for_plotting}
\title{Format Model Outcomes for Plotting}
\usage{
format_simulation_outcomes_for_plotting(sim_outcomes)
}
\description{
Format Simulation Outcomes in a "tidy" (Long & Lean) fashion (as opposed to 
wide, with compartments in columns).
}
\examples{
# setup
param_vec <- load_parameters()
det_table <- make_detection_table(t = 180, rdetecti = .1, rdetecta = 0)

# run simulation
sim_out <- run_param_vec(param_vec, days_out1 = 180, days_out3 = 180, det_table = det_table)

# format data for plotting
sim_out_formatted <- format_simulation_outcomes_for_plotting(sim_out)

# plot! 
plot_flows_by_compartment(sim_out_formatted)

}
