% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_cumulative_cases_by_age_int}
\alias{plot_cumulative_cases_by_age_int}
\title{Plot Cumulative Cases by Age - Intervention}
\usage{
plot_cumulative_cases_by_age_int(out_cases)
}
\description{
Plot Cumulative Cases by Age - Intervention
}
