% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes.R
\name{plot_fit_to_observed_data}
\alias{plot_fit_to_observed_data}
\title{Plot Fit to Observed Data}
\usage{
plot_fit_to_observed_data(out,
  observed_data = read.csv(system.file("time_series/time_series_SCC.csv",
  package = "covid.epi"), as.is = T))
}
\description{
Plot Fit to Observed Data
}
