% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_param_vec}
\alias{run_param_vec}
\title{Run Model from Parameter Vector}
\usage{
run_param_vec(params, params2 = NULL, p.adj = NA, obs.adj = NA,
  days_out1 = 30, days_out2 = NULL, model_type = run_basic)
}
\description{
Run Model from Parameter Vector
}
