% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Define server logic required to run the model and display the results}
\usage{
server(input, output, session)
}
\description{
Define server logic required to run the model and display the results
}
